+++ DawnLib.Dusk/src/API/Config/Weights/Transformers/MoonWeightTransformer.cs
@@ -31,22 +31,38 @@ public class MoonWeightTransformer : WeightTransformer
 
     public override void FromConfigString(string config)
     {
-        if (string.IsNullOrEmpty(config))
-            return;
-
         MatchingMoonsWithWeightAndOperationDict.Clear();
-        IEnumerable<string> configEntries = config.ToLowerInvariant().Split(',', StringSplitOptions.RemoveEmptyEntries);
+        List<string> configEntries = new();
+        foreach (string entry in config.ToLowerInvariant().Split(',', StringSplitOptions.RemoveEmptyEntries))
+        {
+            configEntries.Add(entry.Trim().Replace(" ", "_"));
+        }
         List<string[]> moonWithWeightEntries = configEntries.Select(kvp => kvp.Split('=', StringSplitOptions.RemoveEmptyEntries)).ToList();
+        if (moonWithWeightEntries.Count == 0)
+        {
+            DuskPlugin.Logger.LogWarning($"Invalid moon weight config: {config}");
+            DuskPlugin.Logger.LogWarning($"Expected Format: <Namespace>:<Key>=<Operation><Value> | i.e. magic_wesleysmod:trite=+20");
+            return;
+        }
+
         foreach (string[] moonWithWeightEntry in moonWithWeightEntries)
         {
             if (moonWithWeightEntry.Length != 2)
+            {
+                DuskPlugin.Logger.LogWarning($"Invalid moon weight entry: {string.Join(",", moonWithWeightEntry)} from config: {config}");
+                DuskPlugin.Logger.LogWarning($"Expected Format: <Namespace>:<Key>=<Operation><Value> | i.e. magic_wesleysmod:trite=+20");
                 continue;
+            }
 
             NamespacedKey moonNamespacedKey = NamespacedKey.ForceParse(moonWithWeightEntry[0].Trim());
 
             string weightFactor = moonWithWeightEntry[1].Trim();
             if (string.IsNullOrEmpty(weightFactor))
-                continue;
+            {
+                DuskPlugin.Logger.LogWarning($"Invalid moon weight entry: {string.Join(",", moonWithWeightEntry)} from config: {config}");
+                DuskPlugin.Logger.LogWarning($"Entry did not have a provided weight factor, defaulting to 0.");
+                weightFactor = "+0";
+            }
 
             MatchingMoonsWithWeightAndOperationDict.Add(moonNamespacedKey, weightFactor);
         }