+++ DawnLib/src/API/Tags/VanillaAutoTagger.cs
@@ -1,7 +1,6 @@
 ﻿namespace Dawn;
 public class VanillaAutoTagger<T> : IAutoTagger<T> where T : INamespaced<T>, ITaggable
 {
-
     public NamespacedKey Tag => Tags.Vanilla;
     public bool ShouldApply(T info)
     {